% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{daily_siri_summer}
\alias{daily_siri_summer}
\title{Helper Function for Breeding Season SIRI Simulation}
\usage{
daily_siri_summer(
  Sj_abundance,
  Sa_abundance,
  I1j_abundance,
  I1a_abundance,
  Rj_abundance,
  Ra_abundance,
  I2j_abundance,
  I2a_abundance,
  fecundity,
  transmission_Sj_summer,
  transmission_Sa_summer,
  transmission_Rj_summer,
  transmission_Ra_summer,
  recovery_I1j_summer,
  recovery_I1a_summer,
  recovery_I2j_summer,
  recovery_I2a_summer,
  mortality_Sj_summer,
  mortality_Sa_summer,
  mortality_I1j_summer,
  mortality_I1a_summer,
  mortality_Rj_summer,
  mortality_Ra_summer,
  mortality_I2j_summer,
  mortality_I2a_summer,
  season_length,
  abundance_threshold,
  density_max,
  habitat_suitability
)
}
\arguments{
\item{Sj_abundance}{A matrix of susceptible juvenile abundances.}

\item{Sa_abundance}{A matrix of susceptible adult abundances.}

\item{I1j_abundance}{A matrix of juveniles infected for the first time.}

\item{I1a_abundance}{A matrix of adults infected for the first time.}

\item{Rj_abundance}{A matrix of recovered juveniles.}

\item{Ra_abundance}{A matrix of recovered adults.}

\item{I2j_abundance}{A matrix of juveniles infected for the second time.}

\item{I2a_abundance}{A matrix of adults infected for the second time.}

\item{fecundity}{Seasonal fecundity of adults.}

\item{transmission_Sj_summer}{Transmission rate from susceptible juveniles to first-time infected juveniles in summer.
A single numeric value.}

\item{transmission_Sa_summer}{Transmission rate from susceptible adults to first-time infected adults in summer.
A single numeric value.}

\item{transmission_Rj_summer}{Transmission rate from recovered juveniles to second-time infected juveniles in summer.
A single numeric value.}

\item{transmission_Ra_summer}{Transmission rate from recovered infected adults to second-time infected adults in summer.
A single numeric value.}

\item{recovery_I1j_summer}{Recovery rate from first-time infected juveniles to recovered juveniles in summer.
A single numeric value.}

\item{recovery_I1a_summer}{Recovery rate from first-time infected adults to recovered adults in summer.
A single numeric value.}

\item{recovery_I2j_summer}{Recovery rate from second-time infected juveniles to recovered juveniles in summer.
A single numeric value.}

\item{recovery_I2a_summer}{Recovery rate from second-time infected adults to recovered adults in summer.
A single numeric value.}

\item{mortality_Sj_summer}{Mortality rate of susceptible juveniles in summer.
A single numeric value.}

\item{mortality_Sa_summer}{Mortality rate of susceptible adults in summer.
A single numeric value.}

\item{mortality_I1j_summer}{Mortality rate of first-time infected juveniles in summer.
A single numeric value.}

\item{mortality_I1a_summer}{Mortality rate of first-time infected adults in summer.
A single numeric value.}

\item{mortality_Rj_summer}{Mortality rate of recovered juveniles in summer.
A single numeric value.}

\item{mortality_Ra_summer}{Mortality rate of recovered adults in summer.
A single numeric value.}

\item{mortality_I2j_summer}{Mortality rate of second-time infected juveniles in summer.
A single numeric value.}

\item{mortality_I2a_summer}{Mortality rate of second-time infected adults in summer.
A single numeric value.}

\item{season_length}{A vector of season lengths in days.}

\item{abundance_threshold}{A vector of quasi-extinction thresholds below which a
population becomes extinct.}

\item{density_max}{The maximum population density for a population.}

\item{habitat_suitability}{A numeric vector that indicates the habitat suitabilities
for the populations.}
}
\value{
A matrix of 8 rows by N populations, where N is the length of the
input population matrices.
}
\description{
This function is an internal one for my Frankenmodel (epizootic + metaRange)
which performs daily simulations of disease dynamics and demography for all
populations at one breeding season timestep.
}
