% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{daily_siri_winter}
\alias{daily_siri_winter}
\title{Helper Function for Non-Breeding Season SIRI Simulation}
\usage{
daily_siri_winter(
  Sj_abundance,
  Sa_abundance,
  I1j_abundance,
  I1a_abundance,
  Rj_abundance,
  Ra_abundance,
  I2j_abundance,
  I2a_abundance,
  transmission_Sj_winter,
  transmission_Sa_winter,
  transmission_Rj_winter,
  transmission_Ra_winter,
  recovery_I1j_winter,
  recovery_I1a_winter,
  recovery_I2j_winter,
  recovery_I2a_winter,
  mortality_Sj_winter,
  mortality_Sa_winter,
  mortality_I1j_winter,
  mortality_I1a_winter,
  mortality_Rj_winter,
  mortality_Ra_winter,
  mortality_I2j_winter,
  mortality_I2a_winter,
  season_length,
  abundance_threshold,
  density_max,
  habitat_suitability
)
}
\arguments{
\item{Sj_abundance}{A matrix of susceptible juvenile abundances.}

\item{Sa_abundance}{A matrix of susceptible adult abundances.}

\item{I1j_abundance}{A matrix of juveniles infected for the first time.}

\item{I1a_abundance}{A matrix of adults infected for the first time.}

\item{Rj_abundance}{A matrix of recovered juveniles.}

\item{Ra_abundance}{A matrix of recovered adults.}

\item{I2j_abundance}{A matrix of juveniles infected for the second time.}

\item{I2a_abundance}{A matrix of adults infected for the second time.}

\item{transmission_Sj_winter}{Transmission rate from susceptible juveniles to first-time infected juveniles in winter.
A single numeric value.}

\item{transmission_Sa_winter}{Transmission rate from susceptible adults to first-time infected adults in winter.
A single numeric value.}

\item{transmission_Rj_winter}{Transmission rate from recovered juveniles to second-time infected juveniles in winter.
A single numeric value.}

\item{transmission_Ra_winter}{Transmission rate from recovered infected adults to second-time infected adults in winter.
A single numeric value.}

\item{recovery_I1j_winter}{Recovery rate from first-time infected juveniles to recovered juveniles in winter.
A single numeric value.}

\item{recovery_I1a_winter}{Recovery rate from first-time infected adults to recovered adults in winter.
A single numeric value.}

\item{recovery_I2j_winter}{Recovery rate from second-time infected juveniles to recovered juveniles in winter.
A single numeric value.}

\item{recovery_I2a_winter}{Recovery rate from second-time infected adults to recovered adults in winter.
A single numeric value.}

\item{mortality_Sj_winter}{Mortality rate of susceptible juveniles in winter.
A single numeric value.}

\item{mortality_Sa_winter}{Mortality rate of susceptible adults in winter.
A single numeric value.}

\item{mortality_I1j_winter}{Mortality rate of first-time infected juveniles in winter.
A single numeric value.}

\item{mortality_I1a_winter}{Mortality rate of first-time infected adults in winter.
A single numeric value.}

\item{mortality_Rj_winter}{Mortality rate of recovered juveniles in winter.
A single numeric value.}

\item{mortality_Ra_winter}{Mortality rate of recovered adults in winter.
A single numeric value.}

\item{mortality_I2j_winter}{Mortality rate of second-time infected juveniles in winter.
A single numeric value.}

\item{mortality_I2a_winter}{Mortality rate of second-time infected adults in winter.
A single numeric value.}

\item{season_length}{A vector of season lengths in days.}

\item{abundance_threshold}{A vector of quasi-extinction thresholds below which a
population becomes extinct.}

\item{density_max}{The maximum population density for a population.}

\item{habitat_suitability}{A numeric vector that indicates the habitat suitabilities
for the populations.}
}
\value{
A matrix of 8 rows by N populations, where N is the length of the
input population matrices.
}
\description{
This function is an internal one for my Frankenmodel (epizootic + metaRange)
which performs daily simulations of disease dynamics and demography for all
populations at one non-breeding season timestep.
}
