% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecology.R
\name{dispersal}
\alias{dispersal}
\title{Dispersal process.}
\usage{
dispersal(dispersal_kernel, abundance, suitability)
}
\arguments{
\item{dispersal_kernel}{\verb{<matrix>} dispersal kernel.}

\item{abundance}{\verb{<matrix>} abundance matrix.}

\item{suitability}{\verb{<matrix>}  optional suitability matrix \verb{range: 0, 1}.}
}
\value{
\verb{<matrix>} Dispersed abundance matrix.
}
\description{
Disperse a (abundance) matrix using a dispersal kernel and optional suitability weights.
}
\details{
The abundance matrix is dispersed using the dispersal kernel.
If a suitability matrix is supplied, the individuals will redistribute
within the dispersal kernel according to the suitability matrix.
This means that individuals will move towards more suitable areas,
if they can reach them.
}
\examples{
n <- 10
n2 <- n^2
abu <- matrix(1:n2, nrow = n, ncol = n)
suitab <- matrix(1, nrow = n, ncol = n)
kernel <- calculate_dispersal_kernel(
    max_dispersal_dist = 4,
    kfun = negative_exponential_function,
    mean_dispersal_dist = 1.2
)
res1 <- dispersal(
    dispersal_kernel = kernel,
    abundance = abu
)
res2 <- dispersal(
    dispersal_kernel = kernel,
    abundance = abu,
    suitability = suitab
)
stopifnot(sum(res1) - sum(res2) < 0.01)
# Also, both results should be the same within numirical precision,
# since the suitability is 1 everywhere.
}
