% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dispersal_fixed_unweighted}
\alias{dispersal_fixed_unweighted}
\title{Unweighted and fixed sized dispersal}
\usage{
dispersal_fixed_unweighted(abundance, dispersal_kernel)
}
\arguments{
\item{abundance}{\verb{<numeric matrix>} Values need to be non-negative.}

\item{dispersal_kernel}{\verb{<numeric matrix>} Dispersal kernel. Needs to have an odd size.}
}
\value{
\verb{<numeric matrix>} The new abundance matrix.
}
\description{
Dispersal function that uses a fixed sized kernel that isn't influenced by
external factors. The individuals in each cell are redistributed to the
surrounding cells according to probability defined in the dispersal kernel.
Useful for e.g. wind dispersed plants.
}
\keyword{internal}
