% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dispersal_fixed_weighted}
\alias{dispersal_fixed_weighted}
\title{Weighted and fixed sized dispersal}
\usage{
dispersal_fixed_weighted(abundance, weights, dispersal_kernel)
}
\arguments{
\item{abundance}{\verb{<numeric matrix>} Values need to be non-negative.}

\item{weights}{\verb{<numeric matrix>} Values need to non-missing and between 0 and 1 for the result to make sense.
Needs to have same size as abundance.}

\item{dispersal_kernel}{\verb{<numeric matrix>} Dispersal kernel. Needs to have an odd size.}
}
\value{
\verb{<numeric matrix>} The new abundance matrix.
}
\description{
Dispersal function that uses a fixed sized kernel and weighted dispersal towards areas that have a higher weight.
Use case are e.g. animals that can sense their surroundings.
}
\keyword{internal}
