% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metabolic_scaling}
\alias{metabolic_scaling}
\title{Metabolic scaling}
\usage{
metabolic_scaling(
  normalization_constant,
  scaling_exponent,
  mass,
  temperature,
  E,
  k = 8.617333e-05
)
}
\arguments{
\item{normalization_constant}{\verb{<numeric>} normalization constant.}

\item{scaling_exponent}{\verb{<numeric>} allometric scaling exponent of the mass.}

\item{mass}{\verb{<numeric matrix>}  mean (individual) mass.}

\item{temperature}{\verb{<numeric matrix>} temperature in kelvin (K).}

\item{E}{\verb{<numeric>} Activation energy in electronvolts (eV).}

\item{k}{\verb{<numeric>} Boltzmann's constant (eV / K).}
}
\value{
The scaled parameter.
}
\description{
A function to calculate the metabolic scaling of a parameter.
}
\details{
General notes:

Reproduction rate is generally assumed to scale with an exponent of \code{-1/4}
and an activation energy of \verb{-0.65 eV} (important: it's negative).

Carrying capacity is generally assumed to scale with an exponent of \code{-3/4}
and an activation energy of \verb{0.65 eV} (important: it's positive).
But read: (Brown et. al. 2004; Brown & Sibly, 2012) for an in-depth explanation.

Notes on units:

1 electronvolt = 1.602176634 * 10^-19 Joule

Boltzmann constant 1.380649 * 10^-23 Joule/Kelvin

Boltzmann constant in eV/K = 8.617333e-05 = (1.380649 * 10^-23) / (1.602176634 * 10^-19)
}
\references{
Brown, J.H., Gillooly, J.F., Allen, A.P., Savage, V.M. & West, G.B. (2004).
Toward a Metabolic Theory of Ecology. Ecology, 85, 1771–1789.
\url{doi:10.1890/03-9000}

Brown, J.H. and Sibly, R.M. (2012). The Metabolic Theory of Ecology and Its Central Equation.
In Metabolic Ecology (eds R.M. Sibly, J.H. Brown and A. Kodric-Brown).
\url{doi:10.1002/9781119968535.ch2}
}
