% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{metaRange-package}
\alias{metaRange}
\alias{metaRange-package}
\title{metaRange: Mechanistic and Metabolic Constrained Species Distribution Model}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A platform to build spatially and temporally explicit process-based species distribution models, that can include an arbitrary number of environmental factors, species and processes including metabolic constraints and species interactions. Its focus is simulating populations of one or multiple species in a grid based landscape and studying the meta-population dynamics and emergent patterns that arise from the interaction of species under complex environmental conditions. 'metaRange' provides functions for common ecological processes such as negative exponential, kernel-based dispersal (see Nathan et al. (2012) \doi{10.1093/acprof:oso/9780199608898.003.0015}), calculation of the envionmental suitability based on cardinal values ( Yan & Hunt (1999) \doi{10.1006/anbo.1999.0955}), reproduction in form of an Ricker model (see Ricker (1954) \doi{10.1139/f54-039} and Cabral & Schurr (2010) \doi{10.1111/j.1466-8238.2009.00492.x}), as well as metabolic scaling based on the metabolic theory of ecology (see Brown et al. (2004) \doi{10.1890/03-9000} and Brown & Sibly (2012) \doi{10.1002/9781119968535.ch2}).
}
\references{
tba.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/metaRange/metaRange}
  \item Report bugs at \url{https://github.com/metaRange/metaRange/issues}
}

}
\author{
\strong{Maintainer}: Stefan Fallert \email{stefan.fallert@stud-mail.uni-wuerzburg.de} (\href{https://orcid.org/0000-0002-9939-4219}{ORCID})

Authors:
\itemize{
  \item Lea Li \email{lea.li@stud-mail.uni-wuerzburg.de}
  \item Juliano S. Cabral \email{j.sarmentocabral@bham.ac.uk} (\href{https://orcid.org/0000-0002-0116-220X}{ORCID})
}

Other contributors:
\itemize{
  \item Bavarian Ministry of Science and Arts (bayklif) [funder]
  \item Deutsche Bundesstiftung Umwelt (DBU) [funder]
}

}
\keyword{internal}
