% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.metaRangeVariableStorage}
\alias{print.metaRangeVariableStorage}
\title{Print traits or globals}
\usage{
\method{print}{metaRangeVariableStorage}(x, ...)
}
\arguments{
\item{x}{\verb{<metaRangeVariableStorage>} The object to print.}

\item{...}{\verb{<any>} ignored.}
}
\value{
\verb{<invisible x>}
}
\description{
Print method for species traits and simulation globals.
}
\examples{
sim_env <- terra::sds(terra::rast(vals = 1, nrow = 2, ncol = 2))
names(sim_env) <- "env_01"
test_sim <- metaRangeSimulation$new(source_environment = sim_env)
test_sim$add_species("species_01")
test_sim$add_traits(species = "species_01", a = 1)
print(test_sim$species_01$traits)
test_sim$add_globals(b = 2)
print(test_sim$globals)
}
