% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_verbosity}
\alias{set_verbosity}
\title{Set verbosity of metaRange simulation}
\usage{
set_verbosity(verbose)
}
\arguments{
\item{verbose}{\verb{<integer>} message verbosity (see description).}
}
\value{
\verb{<invisible list>} a list with the previous verbosity setting.
}
\description{
Just a wrapper for \verb{options(metaRange.verbose = [0 | 1 | 2])} but documented.
If \code{0}, metaRange functions will print no messages to the console.
If \code{1}, metaRange functions will print some messages to the console.
If \code{2}, metaRange functions will print many messages to the console.
}
\examples{
set_verbosity(0)
getOption("metaRange.verbose")
}
