% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.metaRangeSimulation}
\alias{summary.metaRangeSimulation}
\title{Summary for metaRange simulation}
\usage{
\method{summary}{metaRangeSimulation}(object, ...)
}
\arguments{
\item{object}{\verb{<metaRangeSimulation>} The \link{metaRangeSimulation} object to summarize.}

\item{...}{\verb{<any>} ignored.}
}
\value{
\verb{<invisible NULL>}
}
\description{
Print a summary of the simulation to the console.
}
\examples{
sim_env <- terra::sds(terra::rast(nrow = 2, ncol = 2))
names(sim_env) <- "env_01"
test_sim <- metaRangeSimulation$new(source_environment = sim_env)
test_sim$add_species("species_01")
summary(test_sim)
}
