% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecology.R
\name{calculate_normalization_constant}
\alias{calculate_normalization_constant}
\title{Normalization constant calculation}
\usage{
calculate_normalization_constant(
  parameter_value,
  scaling_exponent,
  mass,
  reference_temperature,
  E = NULL,
  k = 8.617333e-05
)
}
\arguments{
\item{parameter_value}{\verb{<numeric>} paramter value under the reference temperature.}

\item{scaling_exponent}{\verb{<numeric>} allometric scaling exponent of the mass.}

\item{mass}{\verb{<numeric>}  mean (individual) mass.}

\item{reference_temperature}{\verb{<numeric>} reference temperature in kelvin (K).}

\item{E}{\verb{<numeric>} Activation energy in electronvolts (eV).}

\item{k}{\verb{<numeric>} Boltzmann's constant (eV / K).}
}
\value{
The calculated normalization constant.
}
\description{
Calculates the normalization constant for the metabolic scaling
based on a parameter value under the optimal thermal conditions.
}
\examples{
calculate_normalization_constant(
     parameter_value = 1,
     scaling_exponent = -1 / 4,
     mass = 1,
     reference_temperature = 273.15,
     E = -0.65
)
}
\references{
Brown, J.H., Gillooly, J.F., Allen, A.P., Savage, V.M. & West, G.B. (2004).
Toward a Metabolic Theory of Ecology. Ecology, 85, 1771–1789.

Brown, J.H. and Sibly, R.M. (2012). The Metabolic Theory of Ecology and Its Central Equation.
In Metabolic Ecology (eds R.M. Sibly, J.H. Brown and A. Kodric-Brown).
\url{doi:10.1002/9781119968535.ch2}
}
