% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_suitability}
\alias{calculate_suitability}
\title{Calculate (estimate) environmental suitability}
\usage{
calculate_suitability(vmax, vopt, vmin, venv)
}
\arguments{
\item{vmax}{\verb{<numeric>} upper (i.e. maximum) tolerable value}

\item{vopt}{\verb{<numeric>} optimal (i.e. prefered) value}

\item{vmin}{\verb{<numeric>} lower (i.e. minimum) tolerable value}

\item{venv}{\verb{<numeric>} environmental value for which to calculate the suitability}
}
\value{
\verb{<numeric>} environmental suitability
}
\description{
Calculate / estimate the environmental suitability for a given environmental value,
based on the three "cardinal" values of the species for that environmental niche.
}
\details{
The environmental suitability is calculated based on a beta distribution
after a formula provided by Yan & Hunt (1999) (see references paragraph)
\deqn{suitability = (\frac{V_{max} - V_{env}}{V_{max} - V_{opt}}) * (\frac{V_{env} - V_{min}}{V_{opt} - V_{min}})^{\frac{V_{opt} - V_{min}}{V_{max} - V_{opt}}}}
}
\note{
The original formula by Yan & Hunt was only intended to calculate
the relative daily growth rate of plants in relation to temperature. The abstraction to
use this to A) calculate a niche suitability; and B) use it on other
environmental values than temperature might not be valid. However, the assumption that the
environmental suitability for one niche dimension is highest at one optimal value and
decreases towards the tolerable minumum and maximum values seems reasonable.
}
\references{
Weikai Yan, L.A. Hunt, (1999)
An Equation for Modelling the Temperature Response of Plants using only the Cardinal Temperatures,
Annals of Botany,
Volume 84, Issue 5,
Pages 607-614,
ISSN 0305-7364, \url{doi:10.1006/anbo.1999.0955}
}
