% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ricker_reproduction_model}
\alias{ricker_reproduction_model}
\title{Ricker reproduction model c++ version}
\usage{
ricker_reproduction_model(abundance, reproduction_rate, carrying_capacity)
}
\arguments{
\item{abundance}{matrix .}

\item{reproduction_rate}{matrix.}

\item{carrying_capacity}{matrix.}
}
\description{
Ricker reproduction model c++ version
}
\references{
Cabral, J.S. & Schurr, F.M. (2010) Estimating demographic
models for the range dynamics of plant species.
Global Ecology and Biogeography, 19, 85–97.
\url{doi:10.1111/j.1466-8238.2009.00492.x}

Original model:
Ricker, W.E. (1954) Stock and recruitment.
Journal of the Fisheries Research Board of Canada, 11, 559–623.
\url{doi:10.1139/f54-039}
}
