% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_species.r
\name{save_species}
\alias{save_species}
\title{Save function}
\usage{
save_species(x, traits = NULL, prefix = NULL, path, overwrite = FALSE, ...)
}
\arguments{
\item{x}{\verb{<metaRangeSpecies>} \link{metaRangeSpecies} object.}

\item{traits}{\verb{<character>} NULL or a character vector specifying the trait to save.
If NULL, all traits are saved.}

\item{prefix}{\verb{<string>} prefix for the file names or NULL.}

\item{path}{\verb{<string>}path to the directory where the files are saved.}

\item{overwrite}{\verb{<boolean>} overwrite existing files.}

\item{...}{additional arguments passed to \link[terra:writeRaster]{terra::writeRaster}.}
}
\value{
invisible the paths to the saved files.
}
\description{
Saves the specified traits of a \link{metaRangeSpecies} object.
}
\examples{
sim_env <- terra::sds(terra::rast(vals = 1, nrow = 2, ncol = 2))
names(sim_env) <- "env_01"
test_sim <- metaRangeSimulation$new(source_environment = sim_env)
test_sim$add_species("species_01")
test_sim$add_traits("species_01", trait_01 = matrix(1, nrow = 2, ncol = 2))
res_path <- save_species(
    test_sim$species_01,
    traits = "trait_01",
    prefix = basename(tempfile()),
    path = tempdir()
)
# the following should be TRUE
# but might fail due to floating point errors (that's why we round the values)
identical(
     round(terra::as.matrix(terra::rast(res_path), wide = TRUE)),
     round(test_sim$species_01$traits[["trait_01"]])
)

# cleanup
unlink(res_path)
stopifnot(!file.exists(res_path))
}
